% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{one number}

\item{b}{one number \code{\link[base]{sum}}}
}
\value{
the sum of the two number
}
\description{
Add two numbers testing.
}
\examples{
add(1,2)
\donttest{add(3,4)}
}
