% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_microbiome_97BC.R
\name{beeplot}
\alias{beeplot}
\title{Beeplot}
\usage{
beeplot(
  overweight,
  control,
  species,
  group,
  data,
  main = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{overweight}{data for overweight group}

\item{control}{data for control group}

\item{species}{microbiom species}

\item{group}{group variables}

\item{data}{dataframe}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main2}{main title}
}
\value{
plot
}
\description{
Beeplot
}
