% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_study_function.R
\name{format_ins_select}
\alias{format_ins_select}
\title{Format data for MR analysis}
\usage{
format_ins_select(x)
}
\arguments{
\item{x}{Output from \code{\link[TwoSampleMR]{format_data}}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start). If id.exposure or pval.exposure not present they will be generated.}
}
\value{
A data frame with the same columns as x, but with only the SNPs that are insignificant at p<1E-5 and in LD with each other at r2<0.001 removed.
}
\description{
Format data for MR analysis
}
