% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_study_function.R
\name{MR_custom}
\alias{MR_custom}
\title{MR_custom}
\usage{
MR_custom(exp_dat, out_dat, outcome_gwas_name)
}
\arguments{
\item{exp_dat}{exposure data}

\item{out_dat}{outcome data}

\item{outcome_gwas_name}{name of outcome GWAS}
}
\value{
A list of MR results
}
\description{
MR_custom
}
