% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_microbiome_97BC.R
\name{Shannon_diversity}
\alias{Shannon_diversity}
\title{Shannon_diversity}
\usage{
Shannon_diversity(metaplan_species_all)
}
\arguments{
\item{metaplan_species_all}{data for microbiome species}
}
\value{
plot
}
\description{
Shannon_diversity
}
